Refactored file naming conventions
